parcels-web-app/requirements.txt
Django==3.0.7
bcrypt==3.1.7
argon2-cffi==19.2.0
psycopg2-binary==2.8.6
django-redis==4.12.1
pandas==1.1.3
pytest-django==4.0.0
pytest-mock==3.3.1
python-dotenv==0.15.0
django-crispy-forms-1.11.0